move the translator logic to its own file
